# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenDeX Backend API",
    "version": "0.1.0",
    "summary": "Local control API of the OpenDeX backend (Android multi-window mirroring).",
...
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "contact": {
      "name": "OpenDeX on GitHub",
      "url": "https://github.com/Lordof21/OpenDex"
    }
  },
  "paths": {
...
    "/api/v1/input/touch": {
      "post": {
        "tags": [
          "input"
        ],
        "summary": "Input Touch",
        "description": "Fallback REST touch injection for non-streaming clients.",
        "operationId": "input_touch_api_v1_input_touch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TouchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/input/key": {
      "post": {
        "tags": [
          "input"
        ],
        "summary": "Input Key",
        "description": "Injects key events, typed characters, or desktop shortcuts.",
        "operationId": "input_key_api_v1_input_key_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/keyboard/layout-status": {
      "get": {
        "tags": [
          "input"
        ],
        "summary": "Keyboard Layout Status",
        "description": "Checks whether Turkish/Q layout setup notice has been dismissed.",
        "operationId": "keyboard_layout_status_api_v1_keyboard_layout_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          }
        }
      }
    },
    "/api/v1/keyboard/open-layout-settings": {
      "post": {
        "tags": [
          "input"
        ],
        "summary": "Keyboard Open Layout Settings",
        "description": "Directs phone to physical keyboard configuration intent.",
        "operationId": "keyboard_open_layout_settings_api_v1_keyboard_open_layout_settings_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/keyboard/mark-configured": {
      "post": {
        "tags": [
          "input"
        ],
        "summary": "Keyboard Mark Configured",
        "description": "Marks keyboard configuration as completed for current device.",
        "operationId": "keyboard_mark_configured_api_v1_keyboard_mark_configured_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "title": "ErrorDetail",
        "type": "object",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "array"
              },
              {
                "type": "object"
              }
            ],
            "description": "A human-readable message, a machine code (e.g. `not_supported`), or — for 422 — the list of validation errors."
          }
        },
        "required": [
          "detail"
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeyRequest": {
        "properties": {
          "window_id": {
            "type": "string",
            "title": "Window Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "keycode",
              "char",
              "text",
              "shortcut"
            ],
            "title": "Kind"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "char": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "modifiers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modifiers"
          }
        },
        "type": "object",
        "required": [
          "window_id",
          "kind"
        ],
        "title": "KeyRequest"
      },
...
      "TouchRequest": {
        "properties": {
          "window_id": {
            "type": "string",
            "title": "Window Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "tap",
              "down",
              "move",
              "up",
              "long_press",
              "drag",
              "scroll"
            ],
            "title": "Type"
          },
          "x": {
            "type": "integer",
            "title": "X",
            "default": 0
          },
          "y": {
            "type": "integer",
            "title": "Y",
            "default": 0
          },
          "path": {
            "anyOf": [
              {
                "items": {
                  "prefixItems": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "integer"
                    }
                  ],
                  "type": "array",
                  "maxItems": 2,
                  "minItems": 2
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "long_press_first": {
            "type": "boolean",
            "title": "Long Press First",
            "default": false
          },
          "hscroll": {
            "type": "number",
            "title": "Hscroll",
            "default": 0.0
          },
          "vscroll": {
            "type": "number",
            "title": "Vscroll",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "window_id",
          "type"
        ],
        "title": "TouchRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The per-user API token (`OPENDEX_API_TOKEN` or `~/.opendex/api-token`)."
      },
      "tokenQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "The same token as a query parameter — accepted on GET requests and WebSocket handshakes only."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "The API token is missing or wrong (WebSocket: close code 4401).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "ForeignOrigin": {
        "description": "The request carries an `Origin` that is not one of the app's own origins, or a `Host` that is not a loopback name (WebSocket: close code 4403).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "LengthRequired": {
        "description": "The request body was sent with chunked transfer-encoding and no `Content-Length` (chunked bodies are refused).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The request body is larger than `OPENDEX_MAX_BODY_BYTES` (default 2 MB).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8710",
      "description": "Default local listener (OPENDEX_HTTP_HOST / OPENDEX_HTTP_PORT)"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "tokenQuery": []
    }
  ]
}